requirements.txt
appdirs==1.4.3
dj-database-url==0.4.2
Django==1.10.4
django-extensions==1.7.5
django-filter==1.0.1
django-reversion==2.0.8
django-simple-history==1.8.1
djangorestframework==3.5.3
EasyProcess==0.2.3
envparse==0.2.0
google-api-python-client==1.6.0
gspread==0.6.2
httplib2==0.9.2
nodeenv==1.1.2
nose==1.3.7
oauth2client==4.0.0
packaging==16.8
psycopg2==2.7.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2016.10
PyVirtualDisplay==0.2.1
requests==2.12.5
rsa==3.4.2
selenium==3.4.1
six==1.10.0
unittest-xml-reporting==2.1.0
uritemplate==3.0.0
